+++ ceogestao/src/main/java/br/com/ceoestudos/ceogestao/dao/TurmaDAO.java
@@ -1,13 +1,14 @@
 package br.com.ceoestudos.ceogestao.dao;
 
+import br.com.ceoestudos.ceogestao.model.Curso;
 import br.com.ceoestudos.ceogestao.model.Pessoa;
+import br.com.ceoestudos.ceogestao.model.SituacaoTurma;
 import br.com.ceoestudos.ceogestao.model.Turma;
 import java.util.List;
-import java.util.Set;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 /**
  *
@@ -19,6 +20,12 @@ public class TurmaDAO {
     @PersistenceContext
     private EntityManager em;
     
+    @Autowired
+    private CursoDAO cursoDAO;
+    
+    @Autowired
+    private PessoaDAO pessoaDAO;
+    
     public List<Turma> listarTodos(){
         return em.createQuery("select t from Turma t").getResultList();
     }
@@ -26,27 +33,27 @@ public class TurmaDAO {
     public Turma getById(Long id){
         return em.find(Turma.class, id);
     }
-    @Transactional
+    
     public void atualizar(Turma turma){
-//        Turma turmaBD = getById(turma.getId());
-//        turmaBD.setCurso(turma.getCurso());
-//        turmaBD.setDataInicio(turma.getDataInicio());
-//        turmaBD.setDiasDaSemana(turma.getDiasDaSemana());
-//        turmaBD.setMinutosAula(turma.getMinutosAula());
-//        turmaBD.setObservacoes(turma.getObservacoes());
-//        turmaBD.setQuorumMinimo(turma.getQuorumMinimo());
-//        turmaBD.setSala(turma.getSala());
-//        turmaBD.setSituacao(turma.getSituacao());
         em.merge(turma);
     }
-    @Transactional
+    
     public void adicionar(Turma turma){
+        Curso cursoBD = cursoDAO.getById(turma.getCurso().getId());
+        if(turma.getProfessor()!=null){
+            Pessoa professor = pessoaDAO.getById(turma.getProfessor().getIdentificador());
+            turma.setProfessor(professor);
+        }
+        turma.setCurso(cursoBD);
         em.persist(turma);
     }
     
     public void excluir(Long id){
         Turma turma = getById(id);
-        em.merge(turma);
+        if(turma.getSituacao().equals(SituacaoTurma.EM_ANDAMENTO)){
+            throw new RuntimeException("Não é possível excluir uma turma em andamento");
+        }
+        //turma = em.merge(turma);
         em.remove(turma);
     }
 }
